feat: moving the interpolations and all functions related with the terrain contact inside the class TerrainContact
